Move `index.html` to `public/` to fix and simplify builds
`client/dev` and `client/build` result in 404: "Cannot GET /"
This change fixes that and also simplifies build flow. `index.html` is not compiled in any way anyway (pun not intended).

Is this a good idea? You could argue that `public/` should be empty. We could use a Webpack plugin for copying `index.html` in that case. But fro ma practical standpoint that is unnecessary.