Deploy every branch to a corresponding App Engine version
* Using the Google App Engine provider for Travis CI, this commit enables deployment of each branch
in the hollowverse/hollowverse repo to a dedicated App Engine version with a matching name.
* These version do not currently receive any traffic. The traffic has to be split manually in the
App Engine Console for now.
* A Google Cloud Platform service account key is required to deploy this version. The key has access
to all the resources in the Hollowverse project on GCP.
* The key is encrypted with AES-256-CBC using the `travis encrypt-file` command and can only be
decrypted with the private key stored in Travis secure environment variables, which is not publicly
accessible.
* This setup limits Travis CI role in our build process to be a CD provider only, so there are no
longer a Docker build process or tests to be performed, and the script command does nothing.