Make sure all forwarded URLs end with / to avoid redirects (#137)
This fixes an issue where a request to `hollowverse.com/najwa-karam` (without a trailing slash) gets
redirected to `live.hollowverse.com/najwa-karam/`. The reason for this behavior is that WP sends a
`301` redirect to the canonical URL (which is the one ending with a trailing slash) and CloudFront does not follow the redirects.