Replace usage of Travis CLI with openssl
Travis CLI has a documented issue with multiple encrypted files (https://github.com/travis-ci/travis.rb/issues/239): whenever a new file is encrypted, the key values of a previously encrypted file are overwritten, even if the files are unrelated. So this will cause the decryption of the previous files to fail.

One documented workaround (http://docs.travis-ci.com/user/encrypting-files/#Encrypting-multiple-files) is to create a zipped file containing all the files we wish to encrypt. But that means we will have to re-encrypt every sensitive file every time we update or add a single file. Which means we will need to keep the unencrypted files around on development machines to re-encrypt them later.