Enable Strict Transport Security
According to Google Cloud Platform documentation, the recommended way to redirect to HTTPS is to add the `Strict-Transport-Security` header (HSTS) to all requests. When the browser receives a response with that header set, it will automatically load all subsequent requests on that domain via HTTPS, even if the URL of a link starts with `http://`.

This might be a bit dangerous, because if for any reason our HTTPS set up fails, users won't be able to load
any page **and they won't be able to visit the HTTP version**. This is why I propose enabling this header in incremental steps:

* We start with a small max age value: this value tells the browser when to check again if the HSTS configuration needs to be updated. Using a small value gives the users a chance to return to our HTTP version if SSL fails.
* We only enable HSTS if we `POST` to `/enable-https`. This reduces the chances that a regular user is redirected to HTTPS. We can tests the changes ourselves and decide what to do next.