Switch base image to node:alpine
Alpine is a lightweight Linux distribution designed for use in containers and aims to offer the required tools and packages at the minimum size.

The official Alpine-based Node.js image is 67MB, while the default Debian-based one is a whopping 672MB. Both images provide `yarn` by default, we just need to add `git` to be able to install `@hollowverse/common`.

The hollowverse image built from this branch is 435MB. The same image built from master is 1.02GB. That's 58% smaller!

This should result in faster builds and deployments.