Prevent page jumping to top when JS is slow to load (#278)
React Router triggers a location change event when JS loads, `componentDidMount` on `<ScrollToOnMount />`  was being called. after JS is executed on the client. If scripts are slow to load, the user could have scrolled down the page by the time JS is executed and the page would jump to top.