Do not cache dynamic pages between deployments (#293)
* Do not cache pages between deployments

This PR splits the `X-Hollowverse-Actual-Environment` header to two headers:

* `X-Hollowverse-Actual-Environment-Branch`
* `X-Hollowverse-Actual-Environment-Commit-ID`

The latter is used as a field in the `Vary` HTTP header, so this basically tells browsers not to use cached pages if the commit ID of the environment differs

However, the `Vary` header should only be set on "dynamic paths", like notable person pages. Varying all paths on the commit ID nullifies any value of using reproducible filenames between builds. If `icons.621a6cac01f0fe001b3523f241f44bc6.svg` does not change between builds, we should not prevent the browser from using the cached version.

The `X-Hollowverse-Actual-Environment` is also intentionally split into two headers because only the commit ID is important for `Vary`, if a user moves from `beta` to `master` but the commit ID does not change between the two environments, we should ot prevent the browser from using the cached version either.

Why not `Vary` on the `branch` cookie?

From [Best Practices for Using the Vary Header](https://www.fastly.com/blog/best-practices-using-vary-header/)

> However, sometimes cookies are used for A/B testing purposes, in which case it's a good idea to Vary on a custom header and leave the Cookie header intact. This avoids a lot of additional logic to make sure the Cookie header is left for URLs that need it (and are probably not cacheable).

This may or may not (fix) #287