Experiment with the new React context API
The new context API makes moves context from an experimental feature of React to a first-class feature. I'm particularly interested in how the new API could improve our testing experience. Instead of having to tinker with Jest and Webpack and mock module files (which I really dislike), we could just swap the `value` prop of the `Provider` that wraps the `<App />` component with a mock implementation that we could create in the tests on the fly.

Note: this is using a polyfill for the new API, because the official API has been merged into React but not released yet.