Use GET for API queries
Eventually, when we implement user log-in, we may need to have two separate instances of `GraphQLClient`, one for read operations (which would use `GET` and can be cached) and one for mutations (which would use `POST` and not be cached).  We would need to set the `Authorization` header for both clients.