Add promiseToCancelable helper, use it for AsynComponent
React throws an error when trying to `setState` on a component that has been unmounted. This is the case for `AsynComponent` when the `load` function completes after the async component has been removed from the tree. If the user navigates away from an `Image` that has started loading, the component will be unmounted by the time the load has finished.

Inspired by [this article] on the React blog, I've added a new `promiseToCancelable` function that gets rid of this error in `AsyncComponent`. It uses Node.js built-in `EventEmitter` (that also works in browser) to cancel the promise. Ideally we should use the browser's new standard[`EventTarget`](https://developer.mozilla.org/en-US/docs/Web/API/EventTarget) but that just didn't work in Jest.