Try to fix NotablePersonBody jumping while loading (#434)
Observations:

* It happens on all browsers, even on desktop, so it's not a Safari bug
* It happens only once when an NP page is loading for the first time. So it's most likely in `LodableNotablePerson`
* Looking at the structure of `loading` component in `LoadableNotablePerson` we see that it's not exactly the same as the fully loaded page:

```tsx
// loading
() => <NotablePersonBody />

// fully loaded
() => (
 <div classeName={classes.root}>
   <NotablePersonBody />
   {/* other components */}
 </div>
)
```

So I suspect it's being caused by some missing styles not being applied at the root of the loading component. We need to share some CSS rules between `LoadableNotablePerson` and `NotablePerson` and match the structure by adding a wrapper `<div />`.

It's very hard to test this in development because the styles are applied on the fly by `style-loader` and that causes a flash of unstyled text.